 AssistenteMEI: Agente Inteligente de Notas Fiscais/requirements.txt
streamlit
pandas
langchain
langchain-community
llama-cpp-python
Pillow

